--- server/src/bin/queue.rs
@@ -1,3 +1,0 @@
-use tokio;
-
-fn main() {}
