+++ server/src/core/mod.rs
@@ -5,3 +5,4 @@ pub mod errors;
 pub mod queue;
 pub mod server;
 pub mod store_v1;
+pub mod utils;
